srv6,bpf: rename egress_policies.h to srv6.h
[ upstream commit b09561c949a88d38384a2be101db56290993e870 ]

The only functions left in egress_policies.h are SRv6 related.
Let's rename this to 'srv6.h' and update references to the old file
name.
