+++ bpf/lib/nodeport.h
@@ -14,7 +14,7 @@
 #include "lb.h"
 #include "common.h"
 #include "overloadable.h"
-#include "egress_policies.h"
+#include "srv6.h"
 #include "eps.h"
 #include "conntrack.h"
 #include "csum.h"
