+++ bpf/source_names_to_ids.h
@@ -28,7 +28,7 @@ __source_file_name_to_id(const char *const header_name)
 	/* header files from bpf/lib/ */
 	_strcase_(101, "arp.h");
 	_strcase_(102, "drop.h");
-	_strcase_(103, "egress_policies.h");
+	_strcase_(103, "srv6.h");
 	_strcase_(104, "icmp6.h");
 	_strcase_(105, "nodeport.h");
